[nvos][security] test fixes

+++ ngts/tests_nvos/conftest.py
@@ -31,6 +31,7 @@ from dotted_dict import DottedDict
 from ngts.nvos_tools.ib.opensm.OpenSmTool import OpenSmTool
 from ngts.tests_nvos.general.security.authentication_restrictions.constants import RestrictionsConsts
 from ngts.tests_nvos.general.security.security_test_tools.constants import AuthConsts, AaaConsts
+from ngts.tests_nvos.general.security.security_test_tools.tool_classes.RemoteAaaServerInfo import LdapServerInfo
 from ngts.tests_nvos.system.clock.ClockConsts import ClockConsts
 from ngts.tests_nvos.system.clock.ClockTools import ClockTools
 from infra.tools.sql.connect_to_mssql import ConnectMSSQL
@@ -227,10 +228,11 @@ def clear_config(markers, active_aaa_server=None):
                 logging.info('Failed with remote user engine, try with local dut engine')
                 System().aaa.authentication.unset(apply=True).verify_result()
 
-            logging.info('Remove remote users home directories')
-            remote_usernames = [user.username for user in active_aaa_server.users]
-            for username in remote_usernames:
-                TestToolkit.engines.dut.run_cmd(f'sudo rm -rf /home/{username}')
+            if isinstance(active_aaa_server, LdapServerInfo):
+                logging.info('Remove LDAP users home directories')
+                remote_usernames = [user.username for user in active_aaa_server.users]
+                for username in remote_usernames:
+                    TestToolkit.engines.dut.run_cmd(f'sudo rm -rf /home/{username}')
 
         if 'system_profile_cleanup' in markers:
             clear_system_profile_config()
